画像サイズから表示方向を指定、comicタグは無視

+++ run.sh
@@ -2,9 +2,8 @@
 
 if [ -f venv/bin/activate ]; then
     source venv/bin/activate
-    python3 run.py --filter 1 --dir $1 --exclude-tag 'censored','mosaic censoring','large breasts','huge breasts','cover page','dvd cover','speech bubble','spoken heart','heart','monochrome','greyscale','blank speech bubble','bar censor'
-    cat `ls $1/*.txt` > prompts.txt
-    php ./mk_scns.php nmssis < prompts.txt > scnes.php
-    rm prompts.txt
+    python3 run.py --filter 1 --dir $1 --exclude-tag 'censored','mosaic censoring','large breasts','huge breasts','cover page','dvd cover','speech bubble','spoken heart','heart','monochrome','greyscale','blank speech bubble','bar censor','comic'
+    php ./mk_scns.php $1/*.txt | tee scnes.php
+    test -f prompts.txt && rm prompts.txt
 fi
